chore(transactions-table): Add TransactionsTable component
